refactor:error layer package 정리
