+++ app/src/main/java/com/codesoom/assignment/errors/authentication/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.errors.authentication;
 
 public class InvalidTokenException extends RuntimeException {
     public InvalidTokenException(String token) {
