+++ app/src/main/java/com/codesoom/assignment/errors/login/LoginFailException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.errors.login;
 
 public class LoginFailException extends RuntimeException {
     public LoginFailException(String email) {
