+++ app/src/main/java/com/codesoom/assignment/errors/product/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.errors.product;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(Long id) {
