+++ app/src/main/java/com/codesoom/assignment/errors/user/UserEmailDuplicationException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.errors.user;
 
 public class UserEmailDuplicationException extends RuntimeException {
     public UserEmailDuplicationException(String email) {
