+++ app/src/main/java/com/codesoom/assignment/errors/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.errors.user;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(Long id) {
